requirements.txt
beautifulsoup4
streamlit
supabase
python-dotenv
openai>=1.30.0
requests
